+++ src/lib_shell_services/monitor_services.ml
@@ -100,20 +100,6 @@ module S = struct
     |+ multi_field "chain" Chain_services.chain_arg (fun t -> t#chains)
     |> seal
 
-  let legacy_valid_blocks =
-    Tezos_rpc.Service.get_service
-      ~description:
-        "(Deprecated) Monitor all blocks that are successfully applied by the \
-         node, disregarding whether they were selected as the new head or not."
-      ~query:validated_or_apply_blocks_query
-      ~output:
-        (merge_objs
-           (obj2
-              (req "chain_id" Chain_id.encoding)
-              (req "hash" Block_hash.encoding))
-           Block_header.encoding)
-      Tezos_rpc.Path.(path / "valid_blocks")
-
   let validated_blocks =
     Tezos_rpc.Service.get_service
       ~description:
@@ -198,21 +184,6 @@ end
 open Tezos_rpc.Context
 
 let bootstrapped ctxt = make_streamed_call S.bootstrapped ctxt () () ()
-
-let legacy_valid_blocks ctxt ?(chains = [`Main]) ?(protocols = [])
-    ?(next_protocols = []) () =
-  make_streamed_call
-    S.legacy_valid_blocks
-    ctxt
-    ()
-    (object
-       method chains = chains
-
-       method protocols = protocols
-
-       method next_protocols = next_protocols
-    end)
-    ()
 
 let validated_blocks ctxt ?(chains = [`Main]) ?(protocols = [])
     ?(next_protocols = []) () =